Change datatype for price.

+++ database/migrations/2023_09_22_215414_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->text('desc');
-            $table->double('price');
+            $table->decimal('price', 9, 3);
             $table->timestamps();
         });
     }
